Create newsreader_Okeypress
This part 7/10 of the exercise.
  The o shortcut
  Keyboard keys are identified by key codes. The o key is identified by the key code 111.
  We can use event.which contains which key was pressed. If event.which === 111, then we know the o key has been pressed.
  If the o key is pressed, we should toggle the current article's description.. 

I am able to write the code and have the keypress function as requested but the training app tells me it's not working:
"Oops, try again. It looks like the o key isn't opening a description. Remember to toggle the current article's description inside the event handler. Look back at the instructions for the code."

I can't figure out why I'm getting this error